+++ search/storage/bleve_storage/storage.go
@@ -1,4 +1,4 @@
-package bleve
+package bleve_storage
 
 import (
 	"context"
